Add league selection and data display to last season stats
Added the functionality to select a preferred league from a dropdown menu  in the 'Last Season Stats' component. The league data is retrieved from a JSON file using the newly added 'League Service'. The selected league's data is then displayed in a table in the respective view. This change is aimed to increase the usability and interactivity of the application by allowing users to view varying data according to their preferences.