Add filters for team and role in Last Season Stats
Added two new filters, one for the team and one for the role in the LastSeasonStats interface. A new FiltrationDataInterface has been injected into LastSeasonStats to process the filtering. This interface is implemented by FiltrationDataService. The applied filters are now also included in the redirected URL after changing the league. The blade template for LastSeasonStats was adjusted to include the actual filtering in the dropdowns and to display the new filters values. The changes help the users to filter the Last Season Stats based on team and role.